+++ HISTORY.md
@@ -28,6 +28,7 @@ Based on RocksDB 8.1.1
 * sst_dump: display metaindex_handle and the index_handle's offset and size in footer information (#404).
 * Static Pinning: Set the default for mid-percent capacity threshold in scoped pinning policy to 70 (#689).
 * db_bench: Add support for individual scoped pinning policy parameters (#687).
+* Enable speedb features: Constrain the interface of SharedOptions (make immutable) (#740).
 
 ### Bug Fixes
 * Fix RepeatableThread to work properly with on thread start callback feature (https://github.com/speedb-io/speedb/pull/667).